+++ xtask/src/main.rs
@@ -121,8 +121,12 @@ fn exec_impl(command_with_args: &[&str], capture: bool) -> Result<String> {
 
     match process.status.code() {
         Some(0) => (),
-        Some(c) => errs.push(format!("Status code was {c}")),
-        None => errs.push("No status code (terminated by signal?)"),
+        Some(c) => {
+            errs.push(format!("Status code was {c}"));
+        }
+        None => {
+            errs.push("No status code (terminated by signal?)");
+        }
     };
 
     let stdout = str_or_stash(&process.stdout, &mut errs);